this is a prime number generator i had previously written, it reads and writes to a list of prime numbers, it can find prime numbers under a 1,000,000,000 in about 30 minutes

+++ src/PrimeNumberTest.java
@@ -3,7 +3,6 @@ import static org.junit.Assert.assertEquals;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collection;
-import java.util.List;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -24,11 +23,11 @@ public class PrimeNumberTest {
 	@Parameters
 	public static Collection<Object[]> primeNumbers() {
 		return Arrays.asList(new Object[][] {
-				{2, new ArrayList<Integer>()},
-				{3, new ArrayList<Integer>(Arrays.asList(2))},
-				{5, new ArrayList<Integer>(Arrays.asList(2, 3))},
-				{10, new ArrayList<Integer>(Arrays.asList(2, 3, 5, 7))},
-				{20, new ArrayList<Integer>(Arrays.asList(2,3,5,7,11,13,17,19))},
+				{2, new ArrayList<Long>()},
+				{3, new ArrayList<Long>(Arrays.asList(2L))},
+				{5, new ArrayList<Long>(Arrays.asList(2L, 3L))},
+				{10, new ArrayList<Long>(Arrays.asList(2L, 3L, 5l, 7l))},
+				{20, new ArrayList<Long>(Arrays.asList(2L,3L,5L,7L,11L,13L,17L,19L))},
 
 		});
 	}
